Move filter-related structs to separate file
