+++ crates/forge/src/test_filter.rs
@@ -1,8 +1,8 @@
 use crate::shared_cache::FailedTestsCache;
 use anyhow::Result;
+use forge_runner::filtering::{ExcludeReason, FilterResult, TestCaseFilter, TestCaseIsIgnored};
 use forge_runner::package_tests::TestCase;
 use forge_runner::package_tests::with_config_resolved::TestCaseWithResolvedConfig;
-use forge_runner::{ExcludeReason, FilterResult, TestCaseFilter, TestCaseIsIgnored};
 
 #[derive(Debug, PartialEq)]
 // Specifies what tests should be included
